services/backend/requirements.txt
django==5.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-csp
djoser
python-dateutil

WeasyPrint==52.5
openpyxl
pillow
uvicorn[standard]
gunicorn
mssql-django

#dev
python-dotenv
django-debug-toolbar